User | Remove the document flow from the user app.
Reclaim | The role selected by the user on the Reclaim screen should remain the same role that gets selected.

User | Change the text color to match the user app theme.

User | Settings | Removed badges and XP field

User | The signup screen that opens from the login screen link is not according to the design.

User | Profile | The user profile screen is not according to the design.

User | Login / Signup | The checkbox  privacy policy color is not according to the user theme.

All Role | Login / Signup | The checkbox privacy policy color is not according to the user theme.

Added permission in AndroidManifest.xml

pubspec.yaml updated intl package and added flutter_localization and shared pref.