User | When the user signs up again, the old phone number is already added.
All role | signup | The application should display validation only when the phone number has less than 10 digits. If the phone number is exactly 10 digits or more than 10 digits, no validation should be displayed.

Leisure / Restaurant / Wellness | Profile | When the user does not add a correct social link, the application should display a validation message.

Leisure | Profile | The system should only show one screen in the profile setup. Currently, the restaurant flow is being displayed in the profile setup.

Leisure | The document flow is not according to the Design.

Wellness | The document flow is not according to the Design.

Restaurant | Profile | Business Hours | When the user taps the info icon again, the info card should be hidden.

Did some localization.